Rename `graphics` to `Graphics`
