+++ Kernel/Bootloader/Handoff.cpp
@@ -17,7 +17,7 @@
 #include <Memory/paging.hpp>
 #include <Panic.hpp>
 // Generic devices
-#include <Devices/graphics/console.hpp>
+#include <Devices/Graphics/console.hpp>
 #include <Devices/serial/rs232.hpp>
 // Bootloaders
 #include <multiboot/multiboot2.h>
