+++ Kernel/Devices/Graphics/console.hpp
