+++ Kernel/Devices/Graphics/framebuffer.hpp
