+++ Kernel/Entry.cpp
@@ -21,8 +21,8 @@
 // Memory management & paging
 #include <Memory/paging.hpp>
 // Generic devices
-#include <Devices/graphics/graphics.hpp>
-#include <Devices/graphics/console.hpp>
+#include <Devices/Graphics/graphics.hpp>
+#include <Devices/Graphics/console.hpp>
 #include <Devices/rtc/rtc.hpp>
 #include <Devices/serial/rs232.hpp>
 #include <Devices/spkr/spkr.hpp>
